64 bits port: updated data type to avoid warnings
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@46739 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/idrs/idr_getpos.c
@@ -20,7 +20,7 @@
 
 #include "idr.h"
 
-rt_public int idr_getpos(IDR *idrs)
+rt_public size_t idr_getpos(IDR *idrs)
 {
 	/* Return the current position in the data stream */
